+++ tests/EFCore.SoftAudit.Tests/FakeTimeProvider.cs
@@ -2,7 +2,7 @@ using EFCore.SoftAudit.Interfaces;
 
 namespace EFCore.SoftAudit.Tests;
 
-public class FakeTimeProvider(DateTime utcNow) : ITimeProvider
+public sealed class FakeTimeProvider(DateTime utcNow) : ITimeProvider
 {
     public DateTime UtcNow { get; } = utcNow;
 }