+++ tests/EFCore.SoftAudit.Tests/FakeUserProvider.cs
@@ -2,7 +2,7 @@ using EFCore.SoftAudit.Interfaces;
 
 namespace EFCore.SoftAudit.Tests;
 
-public class FakeUserProvider(string userId) : ICurrentUserProvider
+public sealed class FakeUserProvider(string userId) : ICurrentUserProvider
 {
     public string? GetCurrentUserId() => userId;
 }